+++ src/main/scala/com/jcdecaux/datacorp/spark/workflow/v2/Node.scala
@@ -1,10 +1,16 @@
 package com.jcdecaux.datacorp.spark.workflow.v2
 
-case class Node(id: String) {
+case class Node(name: String, stage: Int, input: Array[String], output: String) {
+
+  def getName: String = {
+    name.split("\\.").last
+  }
 
   def describe(): Unit = {
-    println(s"Node: $id")
-    println("--------------")
-
+    println(s"Node   : $getName")
+    println(s"Stage  : $stage")
+    input.foreach(i => println(s"Input  : $i"))
+    println(s"Output : $output") //
+    println("--------------------------------------")
   }
 }